Sync upstream - Fix all of the warnings in PHP74.

+++ swoole_coroutine_util.c
@@ -782,7 +782,7 @@ static PHP_METHOD(swoole_coroutine_util, fread)
         swAio_init();
     }
 
-    swTrace("fd=%d, offset=%ld, length=%ld", fd, ev.offset, ev.nbytes);
+    swTrace("fd=%d, offset=%jd, length=%ld", fd, (intmax_t) ev.offset, ev.nbytes);
 
     int ret = swAio_dispatch(&ev);
     if (ret < 0)
@@ -863,7 +863,7 @@ static PHP_METHOD(swoole_coroutine_util, fgets)
         swAio_init();
     }
 
-    swTrace("fd=%d, offset=%ld, length=%ld", fd, ev.offset, ev.nbytes);
+    swTrace("fd=%d, offset=%jd, length=%ld", fd, (intmax_t) ev.offset, ev.nbytes);
 
     int ret = swAio_dispatch(&ev);
     if (ret < 0)
@@ -949,7 +949,7 @@ static PHP_METHOD(swoole_coroutine_util, fwrite)
 
     php_swoole_check_aio();
 
-    swTrace("fd=%d, offset=%ld, length=%ld", fd, ev.offset, ev.nbytes);
+    swTrace("fd=%d, offset=%jd, length=%ld", fd, (intmax_t) ev.offset, ev.nbytes);
 
     int ret = swAio_dispatch(&ev);
     if (ret < 0)
